+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# [0.0.1-alpha.4] - 2021-03-21
+
+## Added
+- options for `getTokenWithPopup` (`GetTokenWithPopupOptions` and `PopupConfigOptions`)
+- `Auth0ClientOptions` (align with parameter of `createAuth0Client`)
+- warning for missing mandatory fields
+
+## Changed
+- parameter of `createAuth0Client` from `Auth0CreateOptions` to `Auth0ClientOptions`
+
+## Removed
+- `Auth0CreateOptions`
+
 # [0.0.1-alpha.3] - 2021-03-13
 
 ## Added